+++ src/Service/DatasystemProviderServiceInterface.php
@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BlobBundle\Service;
 
-use Symfony\Component\HttpFoundation\File\File;
-use Symfony\Component\HttpFoundation\Response;
+use Psr\Http\Message\StreamInterface;
 
 interface DatasystemProviderServiceInterface
 {
-    public function saveFile(string $internalBucketId, string $fileId, File $file): void;
+    public function saveFile(string $internalBucketId, string $fileId, \SplFileInfo $file): void;
 
-    public function getBinaryResponse(string $internalBucketId, string $fileId): Response;
+    public function getFileStream(string $internalBucketId, string $fileId): StreamInterface;
 
     public function removeFile(string $internalBucketId, string $fileId): void;
 
@@ -23,12 +22,12 @@ interface DatasystemProviderServiceInterface
     public function getFileHash(string $internalBucketId, string $fileId): string;
 
     /**
-     * Returns true if the file exists, i.e. if getBinaryResponse() would return something.
+     * Returns true if the file exists, i.e., if getFileStream() would return something.
      */
     public function hasFile(string $internalBucketId, string $fileId): bool;
 
     /**
-     * Returns an iterable over all available file IDs. i.e. getFileExists() should
+     * Returns an iterator over all available file IDs, i.e., getFileExists() should
      * return true for all of them.
      *
      * @return iterable<string>